correcting unit test for java 17
In java 17 the accuracy of LocalDateTime is in nanoseconds. The H2 database / SQL Standard by default stores the timestamp with a lower accuracy. This can case problems, if the LocalDateTime is stored and then searched. Since it was stored with a lower accuracy there will be no match!

The API and UI itself demands an accuracy of milliseconds which means that the backend won't cause any trouble as the nanosecond part will be filled with 0 and when stored even ignored.